Move Ethereum events endpoint client to helpers mod
